% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{benchInEnv}
\alias{benchInEnv}
\title{Benchmark multiple expressions in multiple environments}
\usage{
benchInEnv(
  expr,
  envName = envList(),
  rep = 3,
  setup = NULL,
  resultAggregation = NULL
)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envName}{A `character()` vector specifying the environment(s) name(s).
If more than one environment is specified, the expression will
be evaluated in each one.}

\item{rep}{An `integer(1)` specifying the number of repetitions for
each expression. Default is 3.}

\item{setup}{An optional `expression` to be evaluated before the
benchmarked expression. For instance, it can be used to load libraries or
retrieve data.}

\item{resultAggregation}{A function to aggregate the results.
By default no aggregation is done, the results will be returned as a numeric
vectors for each environment.}
}
\value{
A list of execution times for each environment:
\itemize{
  \item A numeric vector of execution times for the evaluated expression
if `length(envName) == 1`.
  \item A list of numeric vectors of execution times for each
environment if `length(envName) > 1`.
}
The result contains the user time (in seconds) for each repetition of
the expression.
}
\description{
Measures the execution time for an R expression across multiple environments.
The function runs the expression multiple times
(controlled by `rep` parameter) and returns the execution times for each run
in the specified environments.
}
\note{
This function is **experimental** and may change in future versions.
Use at your own risk.
}
\examples{
envCreate("my_env", packages = c("digest@0.6.36"))
benchInEnv(Sys.sleep(1), "my_env", rep = 3, resultAggregation = mean)

}
