% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{clearEnv}
\alias{clearEnv}
\title{Clear a specific environment}
\usage{
clearEnv(envName)
}
\arguments{
\item{envName}{A `character(1)` string specifying the name of
the environment.}
}
\description{
Delete a specific environment directory.
}
