% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{copyToEnvs}
\alias{copyToEnvs}
\title{Copy files to environments}
\usage{
copyToEnvs(sourcePath, envNames = listEnvs(), targetPath = "")
}
\arguments{
\item{sourcePath}{A `character(1)` string specifying the path of the
file or directory to copy.}

\item{envNames}{A `character()` vector specifying the environments' names
where the file should be copied. Default is all environments.}

\item{targetPath}{A `character(1)` string specifying the relative path
within each environment where the file should be copied.}
}
\description{
Copy a file or directory into multiple environments.
}
