% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{createEnv}
\alias{createEnv}
\title{Create a new environment}
\usage{
createEnv(envName, packages = NULL, lockfile = NULL)
}
\arguments{
\item{envName}{A `character(1)` string specifying the name of
the environment.}

\item{packages}{A `character()` vector of package names to install
in the environment. Default is NULL.}

\item{lockfile}{A `character()` string specifying the path to a lockfile.
Default is NULL.}
}
\description{
Create an environment directory and installs the
specified packages.
}
\details{
Note that at least of the `packages` and `lockfile` should be present in
order to create the environment.
}
