% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{deleteEnv}
\alias{deleteEnv}
\title{Delete environments}
\usage{
deleteEnv(envName = listEnvs())
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of
the environment(s) to delete. Default is all environments.}
}
\description{
Delete environment(s) directory.
}
