% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envCopyTo}
\alias{envCopyTo}
\title{Copy files to environments}
\usage{
envCopyTo(sourcePath, envName = envList(), targetPath = "")
}
\arguments{
\item{sourcePath}{A `character(1)` string specifying the path of the
file or directory to copy.}

\item{envName}{A `character()` vector specifying the environment(s) name(s)
where the file should be copied. Default is all environments.}

\item{targetPath}{A `character(1)` string specifying the relative path
within each environment where the file should be copied. Default is root of
the environment.}
}
\description{
Copy a file or directory into multiple environments.
}
