% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envInfo}
\alias{envInfo}
\title{Display environment(s) information}
\usage{
envInfo(
  envName = envList(),
  statusInfo = TRUE,
  pkgInfo = FALSE,
  fileInfo = FALSE
)
}
\arguments{
\item{envName}{A `character()` vector specifying the environment(s) name(s).
Default is all environments.}
}
\description{
Display the packages installed in the specified environment(s).
}
