% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envInstallPackage}
\alias{envInstallPackage}
\title{Install new package(s) to environment(s)}
\usage{
envInstallPackage(package, envName = envList(), quiet = FALSE)
}
\arguments{
\item{package}{A `character()` vector of package name(s) to install
in the environment.

To specify the packages you want to install within the environment,
you can use different syntax:

\itemize{
  \item \code{pkg}: install the latest version of the package from CRAN.
  \item \code{pkg@version}: install a specific version of the package
from CRAN.
  \item \code{username/repo}: install the package from GitHub.
     You can also specify the commit with \code{username/repo@commitId}.
  \item \code{bioc::pkg}: install the package from Bioconductor.
}

Note that to install a specific version of a Bioconductor package, it is
recommended to install it via GitHub.

You can find more information on package installation in the
\code{\link[renv:install]{renv::install}} documentation.}

\item{envName}{A `character()` string specifying the name(s) of the
environment(s) for which to update the lockFiles.
Default is all environments.}

\item{quiet}{A `logical()` indicating whether messages should be suppressed.
Default is `FALSE`.}
}
\description{
Install new package(s) to environment(s)
}
