% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envList}
\alias{envList}
\title{List existing environments}
\usage{
envList()
}
\value{
A `character()` vector of environment names.
}
\description{
Return a character vector containing the names of all
existing environments.
}
