% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envRemoveFrom}
\alias{envRemoveFrom}
\title{Remove files from environments}
\usage{
envRemoveFrom(targetPath, envName = envList())
}
\arguments{
\item{targetPath}{A `character(1)` string specifying the relative path
within each environment where the file should be removed.}

\item{envName}{A `character()` vector specifying the environment(s) name(s)
where the file should be removed. Default is all environments.}
}
\description{
Remove a file or directory from multiple environments.
}
