% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockFileManagement.R
\name{exportLockfile}
\alias{exportLockfile}
\title{Export environment lockfile}
\usage{
exportLockfile(envName = listEnvs(), exportPaths = NULL)
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of the
environment(s).}

\item{exportPath}{A `character(1)` string specifying the path where the
lockfile should be exported. Default is an auto-generated path.}
}
\description{
Export an environment's lockfile to a specified location.
}
