% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{listEnvs}
\alias{listEnvs}
\title{List existing environments}
\usage{
listEnvs()
}
\value{
A `character()` vector of environment names.
}
\description{
Return a character vector containing the names of all
existing environments.
}
