% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{notifySuccess}
\alias{notifySuccess}
\title{Show success notification}
\usage{
notifySuccess(msg)
}
\arguments{
\item{msg}{Message to display}
}
\description{
Show success notification
}
\keyword{internal}
