% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{removeFromEnvs}
\alias{removeFromEnvs}
\title{Remove files from environments}
\usage{
removeFromEnvs(targetPath, envNames = listEnvs())
}
\arguments{
\item{targetPath}{A `character(1)` string specifying the relative path
within each environment where the file should be removed.}

\item{envNames}{A `character()` vector specifying the environments' names
where the file should be removed. Default is all environments.}
}
\description{
Remove a file or directory from multiple environments.
}
