% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runInEnvironment}
\alias{runInEnvironment}
\alias{runInEnvironments}
\title{Run expressions in environments}
\usage{
runInEnvironment(expr, env)

runInEnvironments(expr, envs = listEnvironments())
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{env}{A `character(1)` string specifying the environment name.}

\item{envs}{A `character()` vector of environment names.
Default is `listEnvironments()`.}
}
\value{
\itemize{
  \item The result of the evaluated expression for `runInEnvironment`.
  \item A list of results for each environment for `runInEnvironments`.
}
}
\description{
These functions allow execution of R expressions
 within specified environments.
\itemize{
  \item `runInEnvironment(expr, env)`: Runs an expression in a
   specific environment.
  \item `runInEnvironments(expr, envs)`: Runs an expression across
   multiple environments.
}
}
\examples{
createEnvironment("my_env", packages = c("digest@0.6.18"))
runInEnvironment(print("Hello from env"), "my_env")
runInEnvironments(print("Running in multiple envs"))

}
