% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockFileManagement.R
\name{updateLockFile}
\alias{updateLockFile}
\title{Update lockFiles of specified environments}
\usage{
updateLockFile(envName = listEnvs())
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of the
environment(s) for which to update the lockFiles.
Default is all environments.}
}
\description{
Update lockFiles of specified environments
}
