% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{withSpinner}
\alias{withSpinner}
\title{Wrap an expression with a loading spinner}
\usage{
withSpinner(message, expr)
}
\arguments{
\item{message}{Spinner message (string)}

\item{expr}{Expression to evaluate}
}
\description{
Wrap an expression with a loading spinner
}
\keyword{internal}
