% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{deleteEnv}
\alias{deleteEnv}
\title{Delete environments}
\usage{
deleteEnv(envName = listEnvs(), force = FALSE)
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of
the environment(s) to delete. Default is all environments.}

\item{force}{A `logical(1)` specifying whether to skip the confirmation}
}
\description{
Delete environment(s) directory.
}
