% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmentManagement.R
\name{envInfo}
\alias{envInfo}
\title{Display environment(s) information}
\usage{
envInfo(
  envName = envList(),
  statusInfo = TRUE,
  pkgInfo = FALSE,
  fileInfo = FALSE
)
}
\arguments{
\item{envName}{A `character()` vector specifying the environment(s) name(s).
Default is all environments.}

\item{statusInfo}{A `logical(1)` specifying whether to display
status information.}

\item{pkgInfo}{A `logical(1)` specifying whether to display
package information.}

\item{fileInfo}{A `logical(1)` specifying whether to display
file information.}
}
\description{
Display environment status, packages installed, and file tree
 in the specified environment(s).
}
