% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environment_management}
\alias{environment_management}
\alias{createEnv}
\alias{clearEnv}
\alias{clearEnvs}
\alias{listEnvs}
\alias{copyToEnvs}
\alias{removeFromEnvs}
\title{Environment Management Functions}
\usage{
createEnv(envName, packages = NULL, lockfile = NULL)

clearEnv(envName)

clearEnvs()

listEnvs()

copyToEnvs(sourcePath, envs = listEnvs(), targetPath = "")

removeFromEnvs(targetPath, envs = listEnvs())
}
\arguments{
\item{envName}{A `character(1)` string specifying the name of the
environment.}

\item{packages}{A `character()` vector of package names to install in the
environment. Default is NULL.}

\item{lockfile}{A `character()` string specifying the path to a lockfile.
Default is NULL.}

\item{sourcePath}{A `character(1)` string specifying the path of the file
or directory to copy.}

\item{envs}{A `character()` that specify the environments' name in which
the file/directory should be removed. Default is `listEnvs()` to remove the
file/directory in all the environments.}

\item{targetPath}{A `character(1)` string specifying the relative path within
each environment where the file or directory should be removed.}
}
\value{
`listEnvs()`: A `character()` vector that contains
 the environments' names
}
\description{
These functions allow users to create, clear, and list environments
 stored in the ".envs" directory.
}
\details{
Description of functions:
\itemize{
     \item `createEnv()`: Creates an environment directory and
      installs the specified packages or uses the lockfile.
     \item `clearEnv()`: Deletes a specific environment directory.
     \item `clearEnvs()`: Deletes all environment directories.
     \item `listEnvs()`: Returns a character vector containing
      the names of all existing environments.
}
}
\examples{
createEnv("my_first_env", packages = c("digest@0.6.18"))
createEnv("my_second_env", packages = c("digest@0.6.17"))
listEnvs()
clearEnv("my_first_env")
listEnvs()
clearEnvs()

}
