% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{exportLockfile}
\alias{exportLockfile}
\title{Export environment lockfile}
\usage{
exportLockfile(envName, exportPath = .lockFileStorageFromEnv(envName))
}
\arguments{
\item{envName}{A `character(1)` string specifying the name of the
environment.}

\item{exportPath}{A `character(1)` string specifying the path where the
lockfile should be exported. Default is an auto-generated path.}
}
\description{
Export an environment's lockfile to a specified location.
}
