% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockFileManagement.R
\name{lockFileExport}
\alias{lockFileExport}
\title{Export environment lockfile}
\usage{
lockFileExport(envName = envList(), exportPath = NULL)
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of the
environment(s). Default is all environments.}

\item{exportPath}{A `character()` string specifying the path(s) where the
lockfile(s) should be exported. Default is an auto-generated path.}
}
\description{
Export an environment's lockfile to a specified location.
}
\details{
envName and exportPath must have the same length.
}
