% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runInEnv}
\alias{runInEnv}
\alias{runInEnvs}
\title{Run expressions in environments}
\usage{
runInEnv(expr, envName)

runInEnvs(expr, envNames = listEnvs())
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envName}{A `character(1)` string specifying the environment name.}

\item{envNames}{A `character()` vector of environment names.
Default is `listEnvs()`.}
}
\value{
\itemize{
  \item The result of the evaluated expression for `runInEnv`.
  \item A list of results for each environment for `runInEnvs`.
}
}
\description{
These functions allow execution of R expressions
 within specified environments.
\itemize{
  \item `runInEnv(expr, envName)`: Runs an expression in a
   specific environment.
  \item `runInEnvs(expr, envNames)`: Runs an expression across
   multiple environments.
}
}
\examples{
createEnv("my_env", packages = c("digest@0.6.18"))
runInEnv(packageVersion("digest"), "my_env")
runInEnvs(packageVersion("digest"))

}
