% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runInEnv}
\alias{runInEnv}
\title{Run expressions in environments}
\usage{
runInEnv(expr, envName = listEnvs())
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envName}{A `character()` specifying the environment(s) name(s).}
}
\value{
\itemize{
  \item The result of the evaluated expression when length(envName) == 1.
  \item A list of results for each environment when length(envName) > 1.
}
}
\description{
Execute R expressions in multiple environments.
}
\examples{
createEnv("my_env", packages = c("digest@0.6.18"))
runInEnv(packageVersion("digest"), "my_env")
runInEnv(packageVersion("digest"))

}
