% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runInEnv}
\alias{runInEnv}
\title{Run expressions in environments}
\usage{
runInEnv(
  expr,
  envName = envList(),
  parallel = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{expr}{An R expression to be evaluated.}

\item{envName}{A `character()` specifying the environment(s) name(s).}

\item{parallel}{A `logical(1)` indicating whether to run
expressions in parallel. Default is `FALSE`.}

\item{ncores}{An `integer(1)` specifying the number of cores to use
for parallel execution. Default is `parallel::detectCores() - 1`.}
}
\value{
\itemize{
  \item The result of the evaluated expression when length(envName) == 1.
  \item A list of results for each environment when length(envName) > 1.
}
}
\description{
Execute R expressions in multiple environments.
}
\examples{
envCreate("my_env", packages = c("jsonlite@1.9.1"))
runInEnv(packageVersion("jsonlite"), "my_env")
runInEnv(packageVersion("jsonlite"))

}
