% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{updateLockFile}
\alias{updateLockFile}
\title{Update lockFiles of specified environments}
\usage{
updateLockFile(envNames)
}
\arguments{
\item{envName}{A `character()` string specifying the name(s) of the
environment(s) for which to update the lockFiles.}
}
\description{
Update lockFiles of specified environments
}
